moves oauthQueryBuilder.js to /fatsecret
